ASAP-70 Re-enable integration tests
